improve plugin structure
