+++ cognite/neat/plugins/__init__.py
@@ -1,3 +1,4 @@
+from ._data_model import DataModelImporterPlugin, DataModelTransformerPlugin
 from ._manager import get_plugin_manager
 
-__all__ = ["get_plugin_manager"]
+__all__ = ["DataModelImporterPlugin", "DataModelTransformerPlugin", "get_plugin_manager"]
