--- cognite/neat/plugins/data_model/importers/__init__.py
@@ -1,5 +1,0 @@
-from ._base import DataModelImporterPlugin
-
-__all__ = [
-    "DataModelImporterPlugin",
-]
