+++ cognite/neat/session/_plugin.py
@@ -4,8 +4,7 @@ from typing import Any
 
 from cognite.neat.core._issues._base import IssueList
 from cognite.neat.core._utils.reader._base import NeatReader
-from cognite.neat.plugins import get_plugin_manager
-from cognite.neat.plugins.data_model.importers._base import DataModelImporterPlugin
+from cognite.neat.plugins import DataModelImporterPlugin, get_plugin_manager
 from cognite.neat.session._experimental import ExperimentalFlags
 
 from ._state import SessionState
